% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_311.R
\name{get_311}
\alias{get_311}
\title{Get Memphis 311 Data}
\usage{
get_311(...)
}
\arguments{
\item{...}{parameters passed on to query (i.e. department = "code enforcement")}
}
\description{
Get Memphis 311 Data
}
