% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memphis311-package.R
\docType{package}
\name{memphis311-package}
\alias{memphis311}
\alias{memphis311-package}
\title{memphis311: Access Memphis 311 Service Requests}
\description{
Access the public dataset for Memphis 311 service requests.
}
\author{
\strong{Maintainer}: Sarah Johnson \email{sarahviamemphis@gmail.com}

}
\keyword{internal}
