% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_311.R
\name{get_311}
\alias{get_311}
\title{Get Memphis 311 Data}
\usage{
get_311(filepath = NULL, department = NULL, date = NULL, cache = TRUE)
}
\arguments{
\item{filepath}{The filepath to a csv file of Memphis 311 Service Requests. If none is given, will call the data from the API.}

\item{department}{The city department for which to call data.}

\item{date}{The last modified date of the requests in YYYY-MM-DD format.}

\item{cache}{Use cache to store data for faster future use.}
}
\description{
Get Memphis 311 Data
}
