% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_311.R
\name{get_311}
\alias{get_311}
\title{Get Memphis 311 Data}
\usage{
get_311(
  source = "file",
  filepath = NULL,
  department = NULL,
  date = NULL,
  cache = TRUE
)
}
\arguments{
\item{source}{File (the default) or API.
\itemize{
\item \code{"cache"}: open cached data
\item \code{"file"}: read a downloaded csv file of Memphis 311 data.
\item \code{"api"}: call the Memphis 311 API (max 1000 rows).
}}

\item{filepath}{If source = "file" and no cache data exists, the csv filepath}

\item{department}{The city department for which to call data}

\item{date}{The date of the requests}

\item{cache}{Use cache to store data for faster future use.}
}
\description{
Get Memphis 311 Data
}
