% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memphis311-package.R
\docType{package}
\name{memphis311-package}
\alias{memphis311}
\alias{memphis311-package}
\title{memphis311: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
